feat: Add endpoint to retrieve storage server names
This commit adds a new endpoint to the StorageServerController that allows clients to retrieve the names of all storage servers. The `getNames` method in the controller calls the `getStorageServerNames` method in the StorageServerService, which fetches all storage servers from the database and maps them to an array of objects containing the server ID and name. The array is then returned as the response.