feat: Add storage server selection to BackupConfigurationCreate form
This commit adds the functionality to select storage servers in the BackupConfigurationCreate form. It introduces the `storageServerNames`, `storageServers`, and `setStorageServers` props to the `BackupConfigurationBasicForm` component. The `storageServerNames` prop is used to populate the dropdown menu with the names of all available storage servers. The selected storage servers are stored in the `storageServers` state using the `setStorageServers` function. This allows users to choose the appropriate storage servers for their backup configuration.