feat: Add enums for backup configuration and storage server status
This commit adds two new enum classes, `BackupConfigurationStatus` and `StorageServerStatus`, to the `App\Enums` namespace. These enums define the possible status values for backup configurations and storage servers, with `ACTIVE` and `INACTIVE` as the available options. These enums will be used to represent the status of backup configurations and storage servers throughout the application.