feat: Add restore functionality to BackupController
This commit adds a new method called `restore` to the `BackupController` class. The `restore` method takes a request object and an ID as parameters and calls the `restoreBackup` method in the `BackupService` class to restore a backup with the specified ID. If the backup is not found, an exception is thrown. This functionality allows users to restore backups in the application.