refactor: Calculate backup size in BackupService
This commit adds a new method called `calculateSize` to the `CommandBuilder` class in order to calculate the size of a backup. The method takes a string parameter `backupManagerWorkDir` and constructs a command using the `du` command to calculate the size of the specified directory. The calculated size is then stored in the `size` property of each backup in the `BackupService` class. This change improves the accuracy of backup size information and enhances the functionality of the application.

Co-authored-by: [Author Name]