refactor: Add getFreeSpace method to storage server drivers
This commit adds a `getFreeSpace` method to the `AwsS3Driver` and `FileSystemDriver` classes in the `StorageServerDrivers` directory. The method retrieves the available free space on the storage server by executing the appropriate command. This change enhances the functionality of the storage server drivers and allows for better monitoring of storage space usage.