refactor: Add getStorageServerFreeSpace method to StorageServerService
This commit adds a new method, `getStorageServerFreeSpace`, to the `StorageServerService` class in the `app/Services` directory. The method retrieves the available free space on a storage server by executing a command built using the `CommandBuilder` class. This change enhances the functionality of the `StorageServerService` and allows for better monitoring of storage space usage.