refactor: Refresh storage server free space in StorageServerService
This commit adds a new method, `refreshStorageServerFreeSpace`, to the `StorageServerService` class in the `app/Services` directory. The method updates the `total_space_free` property of a `StorageServer` object by calling the `getStorageServerFreeSpace` method and saving the changes. This change enhances the functionality of the `StorageServerService` and ensures that the storage server's free space is always up to date.