refactor: Add CalculateFreeSpaceStorageServerJob
This commit adds the `CalculateFreeSpaceStorageServerJob` class to the `app/Jobs` directory. The job is responsible for calculating the free space on a storage server and updating the `total_space_free` property of the corresponding `StorageServer` object. This change enhances the functionality of the application and ensures accurate tracking of storage server free space.