refactor: Update total_space_free_column in StorageServerView.tsx
This commit updates the `total_space_free_column` in the `StorageServerView.tsx` file. It now checks if `d.total_space_free` is null and displays "Not Calculated" if true, otherwise it formats the value using the `formatBytes` function. This change improves the display of the total space free column in the storage server view.