refactor: Update StorageServerService to calculate and update total_space_free
This commit updates the `StorageServerService.php` file to calculate and update the `total_space_free` property of a `StorageServer` object. The `storeStorageServer` and `updateStorageServer` methods now call the `getStorageServerFreeSpace` method to retrieve the free space of the storage server and assign it to the `total_space_free` property. This change enhances the functionality of the application and ensures accurate tracking of storage server free space.