refactor: Update StorageServerObserver to calculate and update total_space_free
This commit updates the `StorageServerObserver.php` file to calculate and update the `total_space_free` property of a `StorageServer` object. The `created` and `updated` event handlers now call the `getStorageServerFreeSpace` method from the `StorageServerService` class to retrieve the free space of the storage server and assign it to the `total_space_free` property. This change enhances the functionality of the application and ensures accurate tracking of storage server free space.