refactor: Add DataSourceFactory for generating fake data sources
This commit adds a new factory class, DataSourceFactory, for generating fake data sources. The factory defines the default state of the model, including a randomly generated name and a connection configuration with an SSH connection. It also sets the driver configuration to use a file system driver with a specific path. This factory will be used for generating test data in the application.