refactor: Update BackupConfigurationForm to set default scheduleCron value
This commit updates the BackupConfigurationForm component to set a default value for the scheduleCron state variable. Previously, the initial value was an empty string, but now it is set to "0 0 * * *". This change ensures that the backup configuration form always has a valid scheduleCron value, even if it is not provided by the user.