refactor: Add Docker group to sail user
This commit modifies the Dockerfile to add the Docker group to the sail user. By adding the sail user to the Docker group, it allows the user to run Docker commands without needing to use sudo. This improvement enhances the development workflow by simplifying the Docker usage for the sail user.