feat: Add temporary path generator function
The `tmpPathGenerator` function is added to the `CommandBuilder` class to generate a temporary path for backup purposes. This function generates a path in the `/tmp/backup-manager/temp/` directory using a UUID. This addition enhances the functionality of the `CommandBuilder` class by providing a convenient way to generate temporary paths for backups.