chore: Update up.sh script to use default port if not provided
The up.sh script has been modified to prompt the user to enter the desired port for the application. If no port is provided, it will use the default port value. This change improves the user experience by allowing flexibility in configuring the app port when starting the Backup Manager.